Added some functionality to Refinery and moved files to Seal folder
